backend/pyproject.toml
[project]
name = "contract-manager"
version = "0.1.0"
description = "Contract management system"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "strawberry-graphql-django>=0.47",
    "psycopg[binary]>=3.1",
    "django-environ>=0.11",
    "django-auditlog>=3.0",
    "django-cors-headers>=4.3",
    "gunicorn>=21.0",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "pyjwt>=2.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=4.1",
    "ruff>=0.2",
    "mypy>=1.8",
    "django-stubs>=4.2",
    "django-extensions>=3.2",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "DJ"]
ignore = ["E501"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main"]
strict = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "config.settings.local"
